Removed boilerplate from test_rpcs.py and test_rpcs2.py
