+++ examples/helloneighbor/vueclient_rpcs/index.html
@@ -39,6 +39,7 @@
           var vm = this;
 
           // Make a normal RPC call to get_graph().
+//          rpcCall('test_func2')
           rpcCall('get_graph')
           .then(function (response) {
             vm.serverresponse = 'See graph below';
